+++ src/test/java/com/map/gaja/client/apllication/validator/ClientRequestValidatorTest.java
@@ -1,6 +1,6 @@
-package com.map.gaja.client.infrastructure;
+package com.map.gaja.client.apllication.validator;
 
-import com.map.gaja.client.infrastructure.ClientRequestValidator;
+import com.map.gaja.client.apllication.validator.ClientRequestValidator;
 import com.map.gaja.client.infrastructure.file.FileValidator;
 import com.map.gaja.client.infrastructure.file.exception.FileNotAllowedException;
 import com.map.gaja.client.presentation.dto.request.NewClientRequest;
